Moved Profiler views into Profiler folder
