+++ packages/react-devtools-shared/src/devtools/views/Profiler/Tooltip.css
